# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Aamira API", "version": "1.0.0" },
  "servers": [{ "url": "http://localhost:3000/api/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register",
        "requestBody": { "required": true },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": { "required": true },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/orders": {
      "post": {
        "summary": "Create order (auth required)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order",
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}